+++ alarm.py
@@ -12,15 +12,25 @@ import sys
 import datetime
 from playsound import playsound
 
-from PyQt6.QtWidgets import QMainWindow, QPushButton, QApplication, QLabel, QLineEdit, QDialog
-from PyQt6.QtCore import QTimer, Qt
+from PyQt6.QtWidgets import QMainWindow, QPushButton, QApplication, QLabel, QLineEdit, QDialog, QMenu
+from PyQt6.QtCore import QFile, QTimer, Qt, QTextStream
 from PyQt6.QtGui import QFont, QIntValidator, QAction
+
+def _changeStyle(path):
+    app = QApplication.instance()
+    if app is None:
+        return
+
+    print(path)
+    with open(path, 'r') as f:
+        style = f.read()
+        app.setStyleSheet(style)
 
 class TimeDialog(QDialog):
     def __init__(self, parent):
         super().__init__(parent=parent)
         self.setWindowTitle('Set Time')
-        self.setGeometry(100, 100, 300, 170)
+        self.setGeometry(100, 100, 170, 200)
         self.parent = parent
         self._getUI()
     
@@ -33,42 +43,46 @@ class TimeDialog(QDialog):
         self.hoursIn = QLineEdit(self)
         self.hoursIn.move(115, 15)
         self.hoursIn.resize(50, 20)
+
         hoursValidator = QIntValidator(0, 99, self)
         self.hoursIn.setValidator(hoursValidator)
         self.hoursIn.setMaxLength(2)
 
         self.minutesLabel = QLabel(self)
         self.minutesLabel.setText('Minutes (0-59):')
-        self.minutesLabel.setGeometry(10, 30, 200, 50)
+        self.minutesLabel.setGeometry(10, 35, 200, 50)
 
         self.minutesIn = QLineEdit(self)
         self.minutesIn.move(115, 45)
         self.minutesIn.resize(50, 20)
+        
         minsecValidator = QIntValidator(0, 59, self)
         self.minutesIn.setValidator(minsecValidator)
         self.minutesIn.setMaxLength(2)
 
         self.secondsLabel = QLabel(self)
         self.secondsLabel.setText('Seconds (0-59):')
-        self.secondsLabel.setGeometry(10, 60, 200, 50)
+        self.secondsLabel.setGeometry(10, 70, 200, 50)
 
         self.secondsIn = QLineEdit(self)
-        self.secondsIn.move(115, 75)
+        self.secondsIn.move(115, 90)
         self.secondsIn.resize(50, 20)
         self.secondsIn.setValidator(minsecValidator)
         self.secondsIn.setMaxLength(2)
 
         self.cancelButton = QPushButton('Cancel', self)
-        self.cancelButton.setGeometry(65, 125, 80, 40)
+        self.cancelButton.setGeometry(15, 150, 60, 30)
         self.cancelButton.clicked.connect(self._cancelInfo)
 
         self.okButton = QPushButton('OK', self)
-        self.okButton.setGeometry(165, 125, 80, 40)
+        self.okButton.setGeometry(100, 150, 60, 30)
         self.okButton.clicked.connect(self._sendInfo)
 
         self.errorLabel = QLabel(self)
         self.errorLabel.setText('ERROR: Value/s out of range.')
-        self.errorLabel.setGeometry(10, 100, 300, 30)
+        self.errorLabel.setFont(QFont('Arial', 12))
+        self.errorLabel.setStyleSheet("color: red;")
+        self.errorLabel.setGeometry(5, 110, 200, 30)
         self.errorLabel.hide()
 
     def _clearInputs(self):
@@ -125,28 +139,23 @@ class Window(QMainWindow):
         timeMenuAction.triggered.connect(self._getTime)
         menuBar.addAction(timeMenuAction)
 
-        menuBar.addMenu('&Settings')
+        settingsMenu = QMenu("&Settings", self)
+        menuBar.addMenu(settingsMenu)
 
-        self.menuBar().setStyleSheet(
-        """
-        QMenuBar {
-            background-color: #fdfdfd;
-        }
-
-        QMenuBar::item::selected {
-            background-color: #efefef;
-        }
-        """
-        )
+        themesMenu = QMenu('&Themes', self)
+        lightAction = QAction('&Light', self)
+        lightAction.triggered.connect(lambda: _changeStyle("assets/stylesheets/light.qss"))
+        darkAction = QAction('&Dark', self)
+        darkAction.triggered.connect(lambda: _changeStyle("assets/stylesheets/dark.qss"))
+        themesMenu.addAction(lightAction)
+        themesMenu.addAction(darkAction)
+        settingsMenu.addMenu(themesMenu)
     
     def _getUI(self):
         # Variables
         self.count, self.start = 0, False
         self.hours, self.minutes, self.seconds = 0, 0, 0
         self.lastVal = 0
-
-        # Set background color
-        self.setStyleSheet("background-color: #ededed;")
 
         # Create button to set the length of the timer: made obsolete by menu button
         # self.setTimeButton = QPushButton('Set time', self)
@@ -165,25 +174,21 @@ class Window(QMainWindow):
         self.startButton = QPushButton("Start", self)
         self.startButton.setGeometry(15, 130, 60, 40)
         self.startButton.clicked.connect(self._startTimer)
-        self.startButton.setStyleSheet("background-color: #f4f4f4;")
 
         # Create pause button
         self.pauseButton = QPushButton("Pause", self)
         self.pauseButton.setGeometry(85, 130, 60, 40)
         self.pauseButton.clicked.connect(self._pauseTimer)
-        self.pauseButton.setStyleSheet("background-color: #f4f4f4;")
 
         # Create restart button
         self.restartButton = QPushButton("Restart", self)
         self.restartButton.setGeometry(155, 130, 60, 40)
         self.restartButton.clicked.connect(self._restartTimer)
-        self.restartButton.setStyleSheet("background-color: #f4f4f4;")
 
         # Create reset button
         self.resetButton = QPushButton("Reset", self)
         self.resetButton.setGeometry(225, 130, 60, 40)
         self.resetButton.clicked.connect(self._resetTimer)
-        self.resetButton.setStyleSheet("background-color: #f4f4f4;")
 
         # Create test button
         # self.resetButton = QPushButton("Test", self)
@@ -260,4 +265,5 @@ if __name__ == '__main__':
     app = QApplication([])
     win = Window()
     win.show()
+    _changeStyle("assets/stylesheets/light.qss")
     sys.exit(app.exec())